requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SocketIO==5.4.1
eventlet==0.36.2
WTForms==3.1.2
email-validator==2.2.0
python-dotenv==1.0.1
Werkzeug==3.0.6